+++ webapp.py
@@ -12,10 +12,13 @@
 #
 """\
 TX-Pi website.
+
+This is meant to generate static pages, do not run this app on a public server,
+user inputs are not checked.
 """
 import os
 import uuid
-from flask import Flask, render_template
+from flask import Flask, render_template, request
 from flask_misaka import Misaka
 import jinja2
 
@@ -29,7 +32,6 @@ app.config['SECRET_KEY'] = str(uuid.uuid4()).encode('utf-8')
 # Let template rendering fail if Jinja encounters undefined variables
 app.jinja_env.undefined = jinja2.StrictUndefined
 
-
 # Available TX-Pi images
 import configparser
 import json
@@ -40,28 +42,74 @@ _TXPI_IMAGES = json.loads(json.dumps(_TXPI_IMAGES._sections))
 del configparser
 del json
 
+_GITHUB_URL = 'https://github.com/ftCommunity/tx-pi'
 
-@app.route('/en/')
-def home_en():
+
+@app.route('/')
+def index():
     """\
-    English homepage.
+    The root page.
     """
-    return render_template('home_en.html', page_title='TX-Pi')
+    return '''<!DOCTYPE html>
+<html>
+  <head>
+    <script type="text/javascript">
+      var lang = (navigator.language || navigator.userLanguage);
+      if (lang && lang.substr(0,2).toLowerCase() === "de") {
+        window.location = "/de/"
+      }
+    </script>
+    <meta http-equiv="refresh" content="0; url=/en/">
+  </head>
+</html>
+'''
 
 
-@app.route('/en/images/')
-def images_en():
+@app.route('/<lang>/')
+def home(lang):
+    """\
+    Homepage.
+    """
+    return render_template('home_{0}.html'.format(lang))
+
+
+@app.route('/<lang>/images/')
+def images(lang):
     """\
     Renders a page about the images.
     """
     images = _TXPI_IMAGES
-    return render_template('images_en.html', page_title='TX-Pi Images',
-                           images=images)
+    return render_template('images_{0}.html'.format(lang), images=images)
 
 
-@app.route('/en/installation/')
-def installation_en():
+@app.route('/<lang>/installation/')
+def installation(lang):
     """\
     Installation hints.
     """
-    return render_template('installation_en.html', page_title='TX-Pi Installation')
+    return render_template('installation_{0}.html'.format(lang))
+
+
+_MAIN_MENU = (
+    ('Installation', 'installation'),
+    ('Images', 'images'),
+    ('Github', _GITHUB_URL),
+)
+
+@app.context_processor
+def inject_defaults():
+    """\
+    Set some default Jinja variables.
+    """
+    is_en = not request.path.startswith('/de/')
+    # Create a (modifiable) copy of the _MAIN_MENU
+    main_menu = list(_MAIN_MENU)
+    main_menu.append(('Deutsch', '/de/') if is_en else ('English', '/en/'))
+    # Uses upper case names to distinguish them from variables set by the endpoints
+    return { # Variables mainly used for skel.html
+            'LANG': 'en' if is_en else 'de',
+            'MAIN_MENU': main_menu,
+            # Variables used by image.html
+            'RELEASED': 'Released' if is_en else 'Veröffentlicht',
+            'CHECKSUM': 'Checksum (MD5)' if is_en else 'Checksumme (MD5)',
+    }