+++ CHANGES.md
@@ -1,2 +1,5 @@
+# 2019-04-13
+* Added support for German language
+
 # 2019-04-12
 * Initial release